@@ -3,17 +3,17 @@
     "description": "Bundle for loading and generating images",
     "type": "symfony-bundle",
     "require": {
-        "php": "^7.4",
+        "php": "^8.1",
         "ext-curl": "*",
         "ext-mbstring": "*",
         "beberlei/assert": "^3.0",
-        "imagine/imagine": "^1.2",
-        "sensio/framework-extra-bundle": "^5.0",
-        "symfony/config": "^4.4 || ^5.0",
-        "symfony/dependency-injection": "^4.4 || ^5.0",
-        "symfony/http-foundation": "^4.4 || ^5.0",
-        "symfony/http-kernel": "^4.4 || ^5.0",
-        "symfony/options-resolver": "^4.4 || ^5.0"
+        "imagine/imagine": "^1.3",
+        "sensio/framework-extra-bundle": "^5.0 || ^6.0",
+        "symfony/config": "^5.0 || ^6.0",
+        "symfony/dependency-injection": "^5.0 || ^6.0",
+        "symfony/http-foundation": "^5.0 || ^6.0",
+        "symfony/http-kernel": "^5.0 || ^6.0",
+        "symfony/options-resolver": "^5.0 || ^6.0"
     },
     "require-dev": {
         "ergebnis/composer-normalize": "^2.28",
