@@ -14,7 +14,7 @@
     },
     "config": {
         "platform": {
-            "php": "7.4.18"
+            "php": "8.1.8"
         },
         "sort-packages": true,
         "allow-plugins": {
